start combining concats
