+++ src/function/scalar/nested_functions.cpp
@@ -39,7 +39,7 @@ void MapUtil::ReinterpretMap(Vector &result, Vector &input, idx_t count) {
 
 void BuiltinFunctions::RegisterNestedFunctions() {
 	Register<StructExtractFun>();
-	Register<ListConcatFun>();
+//	Register<ListConcatFun>();
 	Register<ListContainsFun>();
 	Register<ListPositionFun>();
 	Register<ListExtractFun>();
