+++ src/include/duckdb/function/scalar/nested_functions.hpp
@@ -89,7 +89,7 @@ struct ListExtractFun {
 
 struct ListConcatFun {
 	static ScalarFunction GetFunction();
-	static void RegisterFunction(BuiltinFunctions &set);
+//	static void RegisterFunction(BuiltinFunctions &set);
 };
 
 struct ListContainsFun {
